Fixed pssg model import when shader >65535 verts

+++ src/EgoEngineLibrary/Formats/Pssg/GltfCarInteriorPssgConverter.cs
@@ -275,9 +275,9 @@ namespace EgoEngineLibrary.Formats.Pssg
                     var b = B + baseVertexIndex;
                     var c = C + baseVertexIndex;
 
-                    rds.Indices.Add((ushort)a);
-                    rds.Indices.Add((ushort)b);
-                    rds.Indices.Add((ushort)c);
+                    rds.Indices.Add((uint)a);
+                    rds.Indices.Add((uint)b);
+                    rds.Indices.Add((uint)c);
                 }
             }
